+++ CHANGELOG.rst
@@ -68,6 +68,7 @@ Added
 Changed
 -------
 
+- Backend defaults values are no longer required (). 
 - The most connected subset in DenseLayout is now reduced bandwidth (#2021).
 - plot_histogram now allows sorting by Hamming distance from target_string (#2064).
 - FunctionalPulse is no longer a class and instead is a decorator, `functional_pulse`